+++ src/contexts/states/index.ts
@@ -4,7 +4,7 @@ import { initBlockState, initBlockListState } from './block'
 import initStatistics from './statistics'
 import initTransactionState from './transaction'
 import initComponents from './components'
-import initNervosDaoState from './nervos_dao'
+import initNervosDaoState from './nervosDao'
 
 export type FetchStatus = keyof State.FetchStatus
